requirements.txt
python_dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
pyyaml